@@ -20,6 +20,6 @@
     }
   ],
   "requirements": {
-    "php": ">=5.5"
+    "php": ">=7.4"
   }
 }
